// ClientApp/apps/general/src/Modules/Shared/Interfaces/IAPIUVGOBannerResponse.ts
export interface IAPIUVGOBannerResponse {
    Id: string;
    EffectiveEndDate: string;
    EffectiveStartDate: string;
    IsEdit: boolean;
    Message: string;
    ApplicationName: string;
    RequestedDate: string;
    NotificationType: string;
    NotificationService: string;
    NotificationTypeId: number;
    NotificationServiceId: number;
    RequesterUserName: string;
    TotalRows: number;
    EmbeddedLink: string;
}